+++ CHANGELOG.markdown
@@ -21,6 +21,7 @@ Improvements:
  * Improve error handling in a couple of places.
  * Use UglifyJS instead of closure compiler to minify JS. It's faster and doesn't require Java :)
  * Update code editor. The new one (CodeMirror) is lighter, faster, smaller and a cleaner implementation.
+ * Add Cmd+Enter / Ctrl+Enter hotkey for saving the document being edited.
  * Prevent "new document" modal from closing on background click.
 
 Bugs: